MAPR-22778 - Return array of KeyValues for AppendRequest
Description -
I think AsyncHbase had a bug where if the user is appending 1 column in the AppendRequest, it returns just 1 KeyValue object, else it throws Exception.
This change fixes that code to always return ArrayList<KeyValue>, or null(If the user did not ask to return results).

Testing Done -
Local unit-testing.

Reviewed By - Aditya